Extract reusable FeedStoreSpecs helper methods into a shared scope. Now can be used by any FeedStore-conformant object
